API: Crud for cluster-volume settings

+++ localpaas_app/db/migrations/20251021060422-create_settings_table.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS settings
     type              VARCHAR(100) NOT NULL,
     kind              VARCHAR(100) NULL,
     name              VARCHAR(100) NULL,
-    size              INT NULL,
+    size              BIGINT NULL,
     version           INT2 NOT NULL DEFAULT 1,
     status            VARCHAR NOT NULL CONSTRAINT chk_status CHECK
                         (status IN ('active','pending','disabled','expired')) DEFAULT 'active',
