+++ localpaas_app/entity/setting.go
@@ -42,7 +42,7 @@ type Setting struct {
 	Kind            string               `bun:",nullzero" json:"kind,omitempty"`
 	Status          base.SettingStatus   `json:"status"`
 	Name            string               `bun:",nullzero" json:"name"`
-	Size            int32                `bun:",nullzero" json:"size"`
+	Size            int64                `bun:",nullzero" json:"size"`
 	Data            string               `bun:",nullzero" json:"data"`
 	AvailInProjects bool                 `json:"availInProjects,omitempty"`
 	Default         bool                 `bun:"is_default" json:"isDefault,omitempty"`
